changes details

+++ src/controllers/cartController.js
@@ -20,21 +20,26 @@ const getCart = async (req, res) => {
   }
 };
 
+let test = false;
 const updateCart = async (req, res) => {
+  if (test) return res.status(200).json({ termino: true });
   try {
+    test = true;
     const { user, carro } = req.body;
 
     const usuario = await User.findOne({ where: { email: user.email }, include: Article });
 
+    console.log("1 solo a la vez");
     await Cartitems.destroy({ where: { userId: usuario.id } });
 
     if (carro.length > 0) {
       carro.forEach(async (element) => {
         const articulo = await Article.findByPk(element.id);
+        console.log("deberia ser 1");
         await usuario.addArticle(articulo, { through: Cartitems, through: { quantity: element.quantity } });
       });
     }
-
+    test = false;
     return res.status(200).json({ termino: true });
   } catch (e) {
     console.log(e);
